llama-chat-backend/requirements.txt
Flask
flask-cors
requests
Pillow 